+++ kotest-property/src/commonMain/kotlin/io/kotest/property/Arbitrary.kt
@@ -1,9 +1,9 @@
 package io.kotest.property
 
-import io.kotest.property.arbitraries.double
-import io.kotest.property.arbitraries.float
-import io.kotest.property.arbitraries.int
-import io.kotest.property.arbitraries.long
+import io.kotest.property.arbitrary.double
+import io.kotest.property.arbitrary.float
+import io.kotest.property.arbitrary.int
+import io.kotest.property.arbitrary.long
 import kotlin.random.Random
 
 /**
