+++ kotest-property/src/commonMain/kotlin/io/kotest/property/arbitrary/doubles.kt
@@ -1,4 +1,4 @@
-package io.kotest.property.arbitraries
+package io.kotest.property.arbitrary
 
 import io.kotest.property.Arbitrary
 import io.kotest.property.PropertyInput
